[bug]: fluid behaviour restored - moved parallelfor into particles struct
