+++ CMakeLists.txt
@@ -27,7 +27,7 @@ find_package(TBB QUIET)
 if(TBB_FOUND)
   message(STATUS "TBB found — using parallel TBB backend")
   target_link_libraries(fluid_core PRIVATE TBB::tbb)
-  target_compile_definitions(fluid_core PRIVATE USE_TBB)
+  target_compile_definitions(fluid_core PUBLIC USE_TBB)
 else()
   message(STATUS "TBB not found — using std::execution backend")
 endif()
@@ -81,7 +81,7 @@ if(BUILD_BENCHMARKS)
 
   if(TBB_FOUND)
     target_link_libraries(fluid-sim-bench PRIVATE TBB::tbb)
-    target_compile_definitions(fluid-sim-bench PRIVATE USE_TBB)
+    target_compile_definitions(fluid-sim-bench PUBLIC USE_TBB)
   endif()
 
   find_package(Threads QUIET)
