+++ src/particle_config.h
@@ -11,7 +11,7 @@ const float INIT_SPACING = 0.014f;          // distance between particles in NDC
 const float INIT_OFFSET_X = -0.3;			// horizontal offset amount in NDC units (negative -> left : positive -> right)
 const float INIT_OFFSET_Y = -0.3;			// vertical offset amount in NDC units (negative -> down : positive -> up)
 const float radius_logical = 2.0f;			// size of particle to be drawn on screen in pixels
-const unsigned int NUM_PARTICLES = 7000;	// number of particles in simulation
+const unsigned int NUM_PARTICLES = 5000;	// number of particles in simulation
 
 // spatial hashing settings
 const float smoothingRadius = 0.05f;	    // defines the area of influence of particle (Increase: slower simulation, greater instability, more accuracy)
@@ -21,9 +21,9 @@ static constexpr int MAX_NEIGHBOURS = 256;  // size of data structure to hold ne
 const float RELAXATION_F = 15000.0f;
 const float ENERGY_RETENTION_F = 0.7f;		// amount of energy retained after colliding with an object (wall, etc.)
 const float MAX_SPEED = 3.0f;				// value to which speed is clamped
-const Vec2  gravity = Vec2{ -0.0f, -3.0f };	// gravity (arbitrary value chosen which looks somewhat realistic)
+const Vec2  gravity = Vec2{ -0.0f, -3.5f };	// gravity (arbitrary value chosen which looks somewhat realistic)
 const float k = 0.000006f;					// scorr coefficient. (Lower = more instability, less damping)
-const int NUM_ITERATIONS = 4;               // number of times to run constraints solver (Lower = greater instability, greater performance)
+const int NUM_ITERATIONS = 3;               // number of times to run constraints solver (Lower = greater instability, greater performance)
 
 // viscocity
 const float xsphC = 0.1f;					// increase for greater viscocity
